Refactor: merge func 'checkSkipStage' into 'shouldSkipStage'
